+++ .vscode/c_cpp_properties.json
@@ -3,7 +3,7 @@
         {
             "name": "Linux",
             "includePath": [
-                "${workspaceFolder}",
+                "${workspaceFolder}/include",
                 "/usr/include/eigen3",
                 "/usr/include/pcl-1.8/**"
             ],
