+++ .vscode/settings.json
@@ -5,6 +5,8 @@
         "vector": "cpp",
         "chrono": "cpp",
         "ostream": "cpp",
-        "iostream": "cpp"
+        "iostream": "cpp",
+        "string_view": "cpp",
+        "tuple": "cpp"
     }
 }
