--- models/env_params.cpp
@@ -1,15 +1,0 @@
-#pragma once
-
-struct EnvParams {
-  int width;
-  int height;
-  double f_xy;
-  int X;
-  int Y;
-  int Z;
-  int roll;
-  int pitch;
-  int yaw;
-
-  bool isFullHeight;
-};
